Refactor_Streaming/src/requirements/development.txt
-r base.txt
django-debug-toolbar>=4.1.0
ipython>=8.12.0
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
factory-boy>=3.2.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.4.0
django-stubs>=4.2.0
pre-commit>=3.3.0
djangorestframework-simplejwt>=5.2.2